Fixed a race condition in turning off the emulator.
Problem: Clicking the `Turn emulator off` link kills the emulator, but the link text does not get updated to `Turn emulator on` on some systems. This was observed in Windows 8.1 x64.

Cause: The link to turn off the emulator kills the emulator process. There's a handler attached on process exit that updates the link text. But because the link handler would immediately dispose the `Process` object after killing it, there may not be sufficient time for the OS to callback into the Exited handler before the object gets disposed.

The fix here delays disposal of the `Process` object to the Exited handler. Note that now the `Process` object will not be disposed anymore upon application exit. But since our application process is being terminated, the OS will release its resources anyway.

+++ src/Controller/UI/Forms/MainForm.cs
@@ -534,21 +534,21 @@ namespace DogAgilityCompetition.Controller.UI.Forms
         {
             this.EnsureOnMainThread(() =>
             {
-                EmulatorProcessKill();
+                if (emulatorProcess != null)
+                {
+                    emulatorProcess.Dispose();
+                    emulatorProcess = null;
+                }
+
                 emulatorLinkLabel.Text = @"Turn emulator on";
             });
         }
 
         private void EmulatorProcessKill()
         {
-            if (emulatorProcess != null)
+            if (emulatorProcess != null && !emulatorProcess.HasExited)
             {
-                if (!emulatorProcess.HasExited)
-                {
-                    emulatorProcess.Kill();
-                }
-                emulatorProcess.Dispose();
-                emulatorProcess = null;
+                emulatorProcess.Kill();
             }
         }
